completed ranking modal functional flow; add anime will be done through this flow; edit ranking db field to be a float; removed api calls to get user's anime list and bookmarks to api.ts

+++ backend/animeranking/models.py
@@ -32,7 +32,7 @@ class Genre(models.Model):
 class UserAnime(models.Model):
     user = models.ForeignKey(UserProfile, on_delete=models.CASCADE)
     anime = models.ForeignKey(Anime, on_delete=models.CASCADE)
-    ranking = models.PositiveIntegerField()
+    ranking = models.FloatField()
 
     class Meta:
         # Ensure a user can't add the same anime more than once
